requirements.txt
absl-py==0.11.0
autopep8==1.5.4
backcall==0.2.0
brotlipy==0.7.0
cachetools==4.2.1
chardet==4.0.0
clyent==1.2.1
codetiming==1.3.0
colorama==0.4.4
common-cmplr-lib-rt==2021.1.2
common-cmplr-lic-rt==2021.1.2
commonmark==0.9.1
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
dpcpp-cpp-rt==2021.1.2
easydict
future==0.18.2
google-auth==1.25.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
importlib-metadata==3.4.0
intel-openmp==2021.1.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==0.14.1
kiwisolver==1.3.1
kornia
lpips
lmdb==1.0.0
loguru==0.6.0
Markdown==3.3.3
matplotlib==3.3.1
memory-profiler==0.58.0
mkl==2021.1.1
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
numpy==1.22.3
oauthlib==3.1.0
opencl-rt==2021.1.2
opencv-python==4.1.2.30
pandas==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.7.3
pyparsing==2.4.7
PyYAML==6.0
requests-oauthlib==1.3.0
rich==9.10.0
rsa==4.7
scipy==1.6.1
tbb==2021.1.1
tensorboard==2.0.0
tensorboard-plugin-wit==1.8.0
thop==0.1.1.post2207130030
toml==0.10.2
torch==1.10.0
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.11.1
torchgeometry
torchmetrics==0.11.3
torch-fidelity
tqdm==4.41.1
ttach==0.0.3
typing-extensions==3.7.4.3
wcwidth==0.2.5
Werkzeug==1.0.1